mlfq: Fix type of time interval

+++ xv6-public/mlfq.h
@@ -2,9 +2,9 @@ struct mlfq {
   int num_queue;
   int sizes[NMLFQ];
   struct proc* queue[NMLFQ][NPROC];
-  unsigned long long quantum[NMLFQ];
-  unsigned long long expire[NMLFQ];
-  unsigned long long boost;
+  uint quantum[NMLFQ];
+  uint expire[NMLFQ];
+  uint boost;
 };
 
 enum mlfqstate {
